+++ tests/Sulu/Component/Rest/ListBuilder/Doctrine/FieldDescriptor/DoctrineCaseFieldDescriptorTest.php
@@ -19,15 +19,17 @@ class DoctrineCaseFieldDescriptorTest extends \PHPUnit_Framework_TestCase
                 'test',
                 new DoctrineDescriptor('entity1', 'field1'),
                 new DoctrineDescriptor('entity2', 'field2'),
-                'CASE WHEN entity1.field1 IS NOT NULL THEN entity1.field1 ELSE entity2.field2 END',
+                '(CASE WHEN entity1.field1 IS NOT NULL THEN entity1.field1 ELSE entity2.field2 END)',
                 [],
+                'entity1.field1 LIKE :search OR (entity1.field1 is NULL AND entity2.field2 LIKE :search)'
             ],
             [
                 'test',
                 new DoctrineDescriptor('test1', 'test2'),
                 new DoctrineDescriptor('test3', 'test4'),
-                'CASE WHEN test1.test2 IS NOT NULL THEN test1.test2 ELSE test3.test4 END',
+                '(CASE WHEN test1.test2 IS NOT NULL THEN test1.test2 ELSE test3.test4 END)',
                 [],
+                'test1.test2 LIKE :search OR (test1.test2 is NULL AND test3.test4 LIKE :search)'
             ],
             [
                 'test',
@@ -42,13 +44,14 @@ class DoctrineCaseFieldDescriptorTest extends \PHPUnit_Framework_TestCase
                     ]
                 ),
                 new DoctrineDescriptor('test3', 'test4'),
-                'CASE WHEN test1.test2 IS NOT NULL THEN test1.test2 ELSE test3.test4 END',
+                '(CASE WHEN test1.test2 IS NOT NULL THEN test1.test2 ELSE test3.test4 END)',
                 [
                     'entity1' => new DoctrineJoinDescriptor(
                         'entity1',
                         'entity.relation'
                     ),
                 ],
+                'test1.test2 LIKE :search OR (test1.test2 is NULL AND test3.test4 LIKE :search)'
             ],
             [
                 'test',
@@ -66,13 +69,14 @@ class DoctrineCaseFieldDescriptorTest extends \PHPUnit_Framework_TestCase
                         ),
                     ]
                 ),
-                'CASE WHEN test1.test2 IS NOT NULL THEN test1.test2 ELSE test3.test4 END',
+                '(CASE WHEN test1.test2 IS NOT NULL THEN test1.test2 ELSE test3.test4 END)',
                 [
                     'entity1' => new DoctrineJoinDescriptor(
                         'entity1',
                         'entity.relation'
                     ),
                 ],
+                'test1.test2 LIKE :search OR (test1.test2 is NULL AND test3.test4 LIKE :search)'
             ],
             [
                 'test',
@@ -96,7 +100,7 @@ class DoctrineCaseFieldDescriptorTest extends \PHPUnit_Framework_TestCase
                         ),
                     ]
                 ),
-                'CASE WHEN test1.test2 IS NOT NULL THEN test1.test2 ELSE test3.test4 END',
+                '(CASE WHEN test1.test2 IS NOT NULL THEN test1.test2 ELSE test3.test4 END)',
                 [
                     'test5' => new DoctrineJoinDescriptor(
                         'test6',
@@ -107,6 +111,7 @@ class DoctrineCaseFieldDescriptorTest extends \PHPUnit_Framework_TestCase
                         'test10'
                     ),
                 ],
+                'test1.test2 LIKE :search OR (test1.test2 is NULL AND test3.test4 LIKE :search)'
             ],
         ];
     }
@@ -114,12 +119,13 @@ class DoctrineCaseFieldDescriptorTest extends \PHPUnit_Framework_TestCase
     /**
      * @dataProvider dataProvider
      */
-    public function testComplete($name, $case1, $case2, $select, $joins)
+    public function testComplete($name, $case1, $case2, $select, $joins, $search)
     {
         $fieldDescriptor = new DoctrineCaseFieldDescriptor($name, $case1, $case2);
 
         $this->assertEquals($select, $fieldDescriptor->getSelect());
         $this->assertEquals($joins, $fieldDescriptor->getJoins());
         $this->assertEquals($name, $fieldDescriptor->getName());
+        $this->assertEquals($search, $fieldDescriptor->getSearch());
     }
 }