+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu
 ==================
 
 * dev-master
+    * HOTFIX      #1962 [ListBuilder]       Fixed search generation with case-field descriptors
     * HOTFIX      #1962 [ContactBundle]     Fixed contact birthday trigger save-button
     * HOTFIX      #1958 [SnippetBundle]     Fixed bug with snippet in snippet